T361 - all block constructors dropped module dependencies
